backend/requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
asgiref==3.11.1
av==17.0.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
ctranslate2==4.7.1
Django==6.0.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
faster-whisper==1.1.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
jellyfish==1.2.1
gTTS==2.5.4
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
mysqlclient==2.2.8
numpy==2.4.3
onnxruntime==1.24.4
openai>=1.0.0
packaging==26.0
pillow==12.1.1
protobuf==7.34.1
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.2.2
PyYAML==6.0.3
RapidFuzz==3.13.0
requests==2.32.3
rich==14.3.3
setuptools==82.0.1
shellingham==1.5.4
sqlparse==0.5.5
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
win32_setctime==1.2.0
django-extensions==3.2.3